# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAnonymi API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/workspaces/{workspace_id}/recovery": {
      "get": {
        "tags": [
          "recovery"
        ],
        "summary": "List Route",
        "operationId": "list_route_api_v1_workspaces__workspace_id__recovery_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workspace Id"
            }
          },
          {
            "name": "document_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecoveryMetadata"
                  },
                  "title": "Response List Route Api V1 Workspaces  Workspace Id  Recovery Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspace_id}/recovery/{snapshot_id}": {
      "get": {
        "tags": [
          "recovery"
        ],
        "summary": "Load Route",
        "operationId": "load_route_api_v1_workspaces__workspace_id__recovery__snapshot_id__get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workspace Id"
            }
          },
          {
            "name": "snapshot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Snapshot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "recovery"
        ],
        "summary": "Save Route",
        "operationId": "save_route_api_v1_workspaces__workspace_id__recovery__snapshot_id__put",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workspace Id"
            }
          },
          {
            "name": "snapshot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Snapshot Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecoveryWrite"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryMetadata"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "recovery"
        ],
        "summary": "Delete Route",
        "operationId": "delete_route_api_v1_workspaces__workspace_id__recovery__snapshot_id__delete",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workspace Id"
            }
          },
          {
            "name": "snapshot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Snapshot Id"
            }
          },
          {
            "name": "expected_version",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Expected Version"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FindingCategory": {
        "type": "string",
        "enum": [
          "person",
          "organization",
          "location",
          "address",
          "identifier",
          "custom",
          "email",
          "phone"
        ],
        "title": "FindingCategory"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RecoveryMetadata": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "workspace_id": {
            "type": "string",
            "format": "uuid",
            "title": "Workspace Id"
          },
          "document_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "workspace_id",
          "document_id",
          "version",
          "updated_at",
          "expires_at"
        ],
        "title": "RecoveryMetadata"
      },
      "RecoveryPayload": {
        "properties": {
          "source": {
            "type": "string",
            "maxLength": 100000,
            "title": "Source"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/FindingCategory"
            },
            "type": "array",
            "maxItems": 8,
            "title": "Categories"
          },
          "phone_region": {
            "type": "string",
            "maxLength": 2,
            "minLength": 2,
            "title": "Phone Region",
            "default": "PH"
          },
          "retention_days": {
            "type": "integer",
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Retention Days",
            "default": 7
          },
          "preset_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preset Id"
          },
          "base_version": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VersionRef"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source"
        ],
        "title": "RecoveryPayload"
      },
...
      "RecoveryView": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "workspace_id": {
            "type": "string",
            "format": "uuid",
            "title": "Workspace Id"
          },
          "document_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "payload": {
            "$ref": "#/components/schemas/RecoveryPayload"
          }
        },
        "type": "object",
        "required": [
          "id",
          "workspace_id",
          "document_id",
          "version",
          "updated_at",
          "expires_at",
          "payload"
        ],
        "title": "RecoveryView"
      },
      "RecoveryWrite": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expected Version"
          },
          "document_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "payload": {
            "$ref": "#/components/schemas/RecoveryPayload"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version",
          "payload"
        ],
        "title": "RecoveryWrite"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VersionRef": {
        "properties": {
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "source_revision_id": {
            "type": "string",
            "format": "uuid",
            "title": "Source Revision Id"
          },
          "decision_version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Decision Version"
          },
          "settings_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Settings Version"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "source_revision_id",
          "decision_version",
          "settings_version"
        ],
        "title": "VersionRef"
      },
...
    }
  }
}